move Qradar transmission methods into qradar_connector
